- move log handler out of PF's package - stub class for PF composition
