+++ folio-service-tools-dev/src/main/java/org/folio/common/log/LogHandler.java
@@ -1,4 +1,4 @@
-package org.folio.common.pf;
+package org.folio.common.log;
 
 @FunctionalInterface
 public interface LogHandler<T> {
